Fix: @Transactional 클래스 분리 [#62] - 기존 트랜잭션이 전체로 적용되어 필요 이상으로 길게 유지되고 롤백 안 되는 문제를 파일 I/O는 트랜지션 밖에서, DB 업데이트만 트랜잭션 안에서 처리하도록 분리
